upgrade

+++ biz/src/main/java/com/hsh24/mall/trade/dao/ITradeDao.java
@@ -16,7 +16,7 @@ public interface ITradeDao {
 	 * @param trade
 	 * @return
 	 */
-	Long createTrade(Trade trade);
+	int createTrade(Trade trade);
 
 	/**
 	 *
