+++ biz/src/main/java/com/hsh24/mall/user/dao/IUserAddressDao.java
@@ -21,7 +21,7 @@ public interface IUserAddressDao {
 	 * @param userAddress
 	 * @return
 	 */
-	Long createUserAddress(UserAddress userAddress);
+	int createUserAddress(UserAddress userAddress);
 
 	/**
 	 *
