+++ biz/src/main/java/com/hsh24/mall/user/dao/IUserWeixinDao.java
@@ -21,6 +21,6 @@ public interface IUserWeixinDao {
 	 * @param user
 	 * @return
 	 */
-	Long createUser(User user);
+	int createUser(User user);
 
 }
